(Integer.MAX_VALUE)

+++ DivisorsOfNum.java
@@ -13,7 +13,7 @@ public class DivisorsOfNum {
 
         int count = 0;
 
-        for (int i = 1; i <= num; i++) {
+        for (long i = 1; i <= num; i++) {
             if (num % i == 0) {
                 System.out.println(i);
                 count++;
